Organize map classes
